feat: refactor SymbolTable and add custom toolbar for symbol scanning functionality
